+++ src/lv_api_map.h
@@ -41,9 +41,39 @@ static inline LV_ATTRIBUTE_TIMER_HANDLER uint32_t lv_task_handler(void)
  *      MACROS
  **********************/
 
- /**********************
-  * DEPRECATED FUNCTIONS
-  **********************/
+
+/**********************
+ * INLINE FUNCTIONS
+ **********************/
+
+/**
+ * Move the object to the foreground.
+ * It will look like if it was created as the last child of its parent.
+ * It also means it can cover any of the siblings.
+ * @param obj       pointer to an object
+ */
+static inline void lv_obj_move_foreground(lv_obj_t* obj)
+{
+    lv_obj_t* parent = lv_obj_get_parent(obj);
+    lv_obj_move_to_index(obj, lv_obj_get_child_cnt(parent) - 1);
+}
+
+/**
+ * Move the object to the background.
+ * It will look like if it was created as the first child of its parent.
+ * It also means any of the siblings can cover the object.
+ * @param obj       pointer to an object
+ */
+static inline void lv_obj_move_background(lv_obj_t* obj)
+{
+    lv_obj_move_to_index(obj, 0);
+}
+
+
+
+/**********************
+ * DEPRECATED FUNCTIONS
+ **********************/
 
 static inline uint32_t lv_obj_get_child_id(const struct _lv_obj_t* obj)
 {