+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v8.1.0 (In progress)
 - renamed lv_obj_get_child_id(obj) to lv_obj_get_index(obj).
 - added lv_obj_move_to_index(obj, index). (#2461)
+- redefined lv_obj_move_foreground(obj) and lv_obj_move_background(obj) as inline functions now calling lv_obj_move_to_index(obj, index).
 - lv_obj_swap(obj1, obj2) added. (#2461)
 - feat(anim) add interface for handling lv_anim user data. (#2415)
 - feat(obj) add lv_is_initialized (#2402)